+++ CHANGELOG.md
@@ -18,8 +18,14 @@ Changelog](http://keepachangelog.com/en/1.0.0/).
 [UNRELEASED](https://github.com/Qiskit/qiskit-aqua/compare/0.7.0...HEAD)
 ========================================================================
 
+Added
+-----
+
+-   Added a `linear` option to the `IsingToQuadraticProgram` converter (#956)
+
 Removed
 -------
+
 -   The monkey-patched ``QuantumCircuit.mcmt`` method and tests, has moved to Terra (#950)
 -   CPLEX as install requirement in ``requirements.txt`` (#952)
 
@@ -107,7 +113,7 @@ Changed
 -   HartreeFock initial state and UCCSD variational form `num_qubits` parameter removed as it was
     only value checked against that computed internally from the other parameters. UCCSD `depth`
     parameter renamed to `reps` and moved in order so it can default to 1. (#939)
-     
+
 
 Removed
 -------